feat: 完善企业招工流程并统一结算与资产金额单位

+++ kuaima_backend/src/main/java/com/kuaima/app/admin/dto/BossOrderView.java
@@ -1,5 +1,7 @@
 package com.kuaima.app.admin.dto;
 
+import java.math.BigDecimal;
+
 /**
  * BossOrder 列表视图：在实体基础上扩展 employerName（雇主企业名/昵称），供 admin 管理后台展示。
  */
@@ -13,7 +15,7 @@ public record BossOrderView(
         String type,
         String postion,
         Integer duration,
-        Integer salary,
+        BigDecimal salary,
         String address,
         String tags,
         String trialDuration,
